+++ CHANGES.md
@@ -1,5 +1,9 @@
 # jtlt CHANGES
 
+## ?
+
+- fix: functional `usePropertySets` argument with DOM joining transformer
+
 ## 0.5.0
 
 - feat: `forEachGroup`